scene file plus non-checked in material for double door.
there was some compiling issue with RAIN scripts missing, which was creating major issues with level meshes and prefabs not showing up, but the recent rollback and discard of local changes seems to have fixed this at least locally. the repo itself may be hosed, but i'm backing things up locally for insurance.